Refactor exegesis_benchmark functionality into a library
This patch refactors the majority of the functionality in the
exegesis_benchmark binary into a library. This allows reuse in other
parts of the code base, particularly beam pipelines once python bindings
have been written.
